Clean up accessibility docs to usable CodeSandbox JS
- remove `ThemeContext` in favor of `useEuiTheme`

+ fix texture example not having correct dark mode theme

+ misc cleanup
