+++ src-docs/src/views/elastic_charts/data.js
@@ -502,29 +502,3 @@ export const DAYS_OF_RAIN = [
     precipitation: 'none',
   },
 ];
-
-export const SAMPLE_SMALL_DATA = [
-  { x: 0, y: 10.934269 },
-  { x: 1, y: 14.660253928 },
-  { x: 2, y: 14.039862184 },
-  { x: 3, y: 13.388734576 },
-  { x: 4, y: 13.051608256 },
-  { x: 5, y: 12.078351999999999 },
-  { x: 6, y: 11.823124744 },
-  { x: 7, y: 11.7314602 },
-  { x: 8, y: 12.89177344 },
-  { x: 9, y: 11.837041216 },
-];
-
-export const SAMPLE_SMALL_DATA_2 = [
-  { x: 0, y: 10.009588 },
-  { x: 1, y: 13.591318 },
-  { x: 2, y: 14.493893151999998 },
-  { x: 3, y: 11.112153736 },
-  { x: 4, y: 12.202770376 },
-  { x: 5, y: 14.948682999999999 },
-  { x: 6, y: 11.527410016000001 },
-  { x: 7, y: 14.404530544 },
-  { x: 8, y: 14.007384976 },
-  { x: 9, y: 14.639346376 },
-];
